Homework 3 has done
